# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenTogg API",
        "description": "Time tracking REST API. Authenticate with a Bearer token created in Settings.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/vectors": {
            "get": {
                "tags": [
                    "Vectors"
                ],
                "summary": "List all vectors",
                "operationId": "1fae4fe35ac174e13dfda3e02c68c6f3",
                "responses": {
                    "200": {
                        "description": "List of vectors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vectors"
                ],
                "summary": "Create a vector",
                "operationId": "e3059ddc85dffc5939f8c4c2eba2dcab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Work"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#BB86FC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vector created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vectors/{id}": {
            "get": {
                "tags": [
                    "Vectors"
                ],
                "summary": "Get a vector",
                "operationId": "401506b6856fbe2ca1c2402b23c31e34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vector external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vector details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vectors"
                ],
                "summary": "Update a vector",
                "operationId": "deebcf69141015daa3f65ec949a97ead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vector external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#BB86FC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vector updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vectors"
                ],
                "summary": "Delete a vector",
                "operationId": "b43073e6b31bef5b7dc192626f220559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Vector external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vector deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}